adversarial_example_generator/requirements.txt
kafka
adversarial-robustness-toolbox
numpy
keras
tensorflow
pillow
requests

